refactor(S-3.01): extract handle_list to auth/list.rs (BC-7.4.013 JSON path preserved)
Move render_list_table, render_list_json, and handle_list out of mod.rs.
render_list_table and render_list_json re-exported under #[cfg(test)] for
snapshot tests in mod.rs. BC-7.4.013 JSON output path preserved.
612 unit tests pass.
